stop faking spark-master name
previously the master would use /tmp/hosts to pretend it was called
spark-master, and workers would likewise to talk to the master.

instead, the name of the master service is passed to the master's
start.sh, and the master pretends to be the service. the service was
already passed to the worker's start.sh and it still is, except now
the worker just tries to connect to the service.