Copy local spark configuration to $SPARK_HOME/conf
Rather than have the pod creator set the SPARK_CONF_DIR environment
variable to /etc/oshinko-spark-configs to modify the spark configuration,
this change simply copies the contents of /etc/oshinko-spark-configs
if present to $SPARK_HOME/conf. This gives us a simple overwrite semantic
so that you only need to specify the files being changed as opposed
to also including copies of files not being changed.