move nss_wrapper setup code to an entrypoint
the goal is to allow users to shell into the cluster pods, or start a
container manually, without having to manually setup nss_wrapper in
their environment.

everyone stumbles on getting the nss_wrapper env setup the first
time. the resulting error is about failure to login from within the
hadoop code, which is confusing. everyone then struggles with the
manual nature of setting up the env after they overcome the strange
login error.