Add signal handlers for launch.sh
In order to terminate spark pods quickly, we need
signal handlers for SIGTERM and SIGINT. Additionally,
we should run the spark processes in the background and
use "wait", recording the PID so that we can issue an
immediate kill to spark.

This is the same strategy we use in the driver.