Cleanup objects in the project after each test in run.sh
This allows multiple test files in the same directory to
run using the same project. A new project will still be
created for each directory level.

Also adds a default value for OPENSHIFT_SPARK_TEST_IMAGE in
run.sh so that test/run.sh can be run outside of make.