Modify entryponit to support spark-on-k8s execution
This change modifies the entrypoint so that if a spark-on-k8s
command is passed, it delegates to the spark entrypoint. Otherwise
the command is executed as is. This allows the openshift-spark
image to continue to function as a Spark Standalone component,
with custom configuration support etc, but double as an OpenShift
spark-on-k8s image.