update change-yaml.sh to contain sha512 logic
Due to upstream changes in the distribution of md5 sum files,
change-yaml.sh will now download the sha512 sum files and the tar
archive to calculate a checksum for validation, then calculate an md5
sum for the cekit schema file. This behavior will be removed as soon as
sha512 support is added in the cekit project.