Allow a modify-spark directory for image completion
This change supports an optional modify-spark directory that
can be put in the build input directory for image completion.
If the modify-spark directory is there, the image completion
script will run rsync to copy it's contents into the spark
install directory. This mechanism allows a user to modify
a spark distro by laying out a parallel directory structure.