Change modify-spark behavior for incomplete images and move chmod
The rsync command used to enable the "modify-spark" behavior
in incomplete images was preserving permissions. However, it
was preserving permissions on files as uploaded by s2i, which
can't be relied on (it's not necessarily the same as perms on
disk from build_input). Therefore, preserving permissiosn has been
dropped, and a chmod to add group rwX to the spark installation
was moved until after the rsync.